requirements.txt
pony == 0.7.14
discord.py == 1.6.0
PyYAML == 5.4.1